Fix code formatting after migration to clang-format v18
* cord/tests/de.c (line_map, history): Define struct type outside
typedef.
* include/private/pthread_support.h (GC_stack_context_t, GC_thread):
Likewise.
* include/private/thread_local_alloc.h (GC_tlfs): Likewise.
* finalize.c (HASH3): Cast 1 to size_t.
* include/gc/cord_pos.h (CORD_pos_advance): Likewise.
* include/gc/gc.h [GC_INITIAL_HEAP_SIZE && !CPPCHECK]
(GC_INIT_CONF_INITIAL_HEAP_SIZE): Cast `GC_INITIAL_HEAP_SIZE` to
`size_t`.
* include/gc/gc_inline.h (GC_FAST_MALLOC_GRANS, GC_MALLOC_WORDS_KIND):
Replace `GC_GRANULE_BYTES*lg` to `GC_RAW_BYTES_FROM_INDEX(lg)`.
* include/private/dbg_mlc.h [(KEEP_BACK_PTRS || MAKE_BACK_GRAPH)
&& !(PARALLEL_MARK && KEEP_BACK_PTRS)] (GC_HAS_DEBUG_INFO): Wrap
`*(GC_uintptr_t*)(base)` into parentheses.
* include/private/gc_priv.h (GRANULES_TO_BYTES, GRANULES_TO_PTRS):
Reverse order of multiplication operands.
* include/private/gc_priv.h [!MARK_BIT_PER_OBJ] (FINAL_MARK_BIT):
Likewise.
* include/private/gc_priv.h [E2K && THREADS] (PS_COMPUTE_ADJUSTED_OFS):
Cast `PS_SYSCALL_TAIL_BYTES` to `unsigned`.
* include/private/gc_priv.h [!_MSC_VER && static_assert && !CPPCHECK
&& __STDC_VERSION__>=201112L] (GC_STATIC_ASSERT): Wrap `static_assert`
into do-while.
* include/private/gcconfig.h [!GC_CLANG_PREREQ(11,0)] (PTR_ALIGN_DOWN,
PTR_ALIGN_UP): Replace `(GC_uintptr_t)((b)-1))` to
`((GC_uintptr_t)(b)-(GC_uintptr_t)1)`.
* include/private/gcconfig.h [I386 && DGUX || X86_64 && MSWIN_XBOX1]
(MAP_FAILED): Replace -1 to `~(GC_uintptr_t)0`.
* win32_threads.c [PARALLEL_MARK && !GC_PTHREADS_PARAMARK && !MSWINCE
&& !MSWIN_XBOX1] (GC_start_mark_threads_inner): Likewise.
* include/private/gcconfig.h [WEBASSEMBLY && WASI] (DATASTART,
DATAEND): Wrap `&` operation into parentheses.
* include/private/gcconfig.h [WEBASSEMBLY && WASI] (STACKBOTTOM):
Change definition to `DATASTART`.
* include/private/gcconfig.h [!FIXUP_POINTER && DYNAMIC_POINTER_MASK]
(FIXUP_POINTER): Wrap `(word)(p)` into parentheses.
* os_dep.c [LINUX && USE_LIBC_PRIVATES && IA64] (os_main_stackbottom):
Move comment into if-else branch; refine comment.
* tests/gctest.c [GC_GCJ_SUPPORT] (gcj_cons): Replace ternary operator
with `if` statement.
* tests/gctest.c (NUMBER_ROUND_UP): Cast 1 to unsigned.
* tests/trace.c (root): Make variable static; define struct type
(`treenode`) before the variable definition.
* win32_threads.c [MSWINCE] (GC_wince_evaluate_stack_min): Cast 1 to
`word`.

+++ finalize.c
@@ -27,7 +27,7 @@ typedef void (*finalization_mark_proc)(ptr_t /* finalizable_obj_ptr */);
 
 #  define HASH3(addr, size, log_size)                               \
     ((size_t)((ADDR(addr) >> 3) ^ (ADDR(addr) >> (3 + (log_size)))) \
-     & ((size)-1))
+     & ((size) - (size_t)1))
 #  define HASH2(addr, log_size) HASH3(addr, (size_t)1 << (log_size), log_size)
 
 struct hash_chain_entry {
